fixed a sonarcube major bug: casted some variable to long in some subtraction

+++ src/main/java/mampf/order/MampfCart.java
@@ -62,7 +62,7 @@ public class MampfCart{
 			removeFromCart(cartItem);
 		}else {
 			domainCart.addOrUpdateItem(cartItem.getProduct(),
-									   itemAmount-cartItem.getQuantity().getAmount().intValue());
+					(long) itemAmount-cartItem.getQuantity().getAmount().intValue());
 		}
 	
 	}
